+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.3.1] - 2020-06-17
+- fix exception in backend when template not found
+- fix wrong return value notation in VideoProviderCollection::getVideoByRawDataWithSelector()
+
 ## [0.3.0] - 2020-06-16
 - made BeforeRenderPlayer event usable
 - added AfterRenderPlayer event